Add meta-data to output

Function to add specific meta data to output. No arguments.

add_meta(...)

Value

Returns a list of the above components

Details

This function captures a specific set of metadata associated with the output. Specifically it captures:

name of the data object used in the run (NB: not the data itself!)

date of run

data.year = most recent year of available NWFSC Observer data. Usually 1 year behind current year.

fishery

state

sector

gear

area

depth

strata = a concatenation of fishery, sector, gear, area, and depth

confidential strata

caption - table caption associated with the table

notes - any notes added when the function was run

Typical use will be inside another function. To enable add_meta to access variables created inside the enclosing function, need to make the execution environment of the enclosing function an object and assign that to the enclosing environment of add_meta.

Examples

if (FALSE) {

xyz_phlb_bycatcher <- function(...) {
  # a bunch of code for the bycatcher...
  ...

  # capture the execution environment of the bycatcher function

  curenv <- new.env()

  # set the enclosing environment of add_meta to be the execution environment
  # the of bycatcher function

  environment(add_meta) <- curenv

  # now add_meta can capture the bycatcher variables

  meta <- add_meta()
} # close bycatcher function
}