Estimate P. halibut Mortality Using Injuries on Hook & Line Vessels

est_mort_injury uses injury assessment data from P. halibut (PHLB) biological specimens (a.k.a., PHLB bio-specimens; NWFSC FOS observer data) collected on hook & line vessels to estimate P. halibut mortality for hook & line fisheries that use injury determinations in the mortality estimation process.

est_mort_injury(
  .df,
  .td,
  .mask = FALSE,
  .confs = NULL,
  path = "~/observer/Input/PHLB/phlbbycatcher/data",
  nmfix = FALSE
)

Arguments

.df
A data.frame containing the PHLB bio-specimen data, where the measured lengths of each specimen have been converted to weight using the International Pacific Halibut Commission length-weight conversion table. The weights have been summarized by viability, year, sector, gear and any other group/strata variables (e.g., area, depth). Typically, this is the output from prep_viabs(..., .mask = FALSE, .confs = NULL).

.td
A data.frame containing the weight of total discards of P. halibut divvied up by year, sector, gear, and any other group/strata. The grouping/stratification must match those found in .df. Typically, these data come from a call to estimate_unsampld_cs(..., .mask = FALSE, .confs = NULL) or similar function.

.mask
Do you want to mask confidential data? Default is *FALSE*. If set to true, must supply a data.frame to the .confs argument. See mask_conf_data().

.confs
(optional) A data.frame containing the unique strata/groups whose accompanying data will be masked. Must be supplied if .mask = TRUE, otherwise, leave as NULL (default).

path
a path to the data/Constants.R file which contains the mortality rates. Default is set to ~/observer/Input/PHLB/phlbbycatcher/.

nmfix
Do you want to fix the names (i.e., namefix()) on output? Default = FALSE.

Value

Returns a data.frame that contains the grouping/stratification variables, and estimates of the weight of P. halibut mortality by injury category, with and without mortality rates applied.

Details

est_mort_injury() will estimate P. halibut mortality for each level of grouping/stratification using the injury data. Thus, this should only be used for those hook & line sectors that use injury determinations in the estimation of P. halibut mortality, which include: IFQ hook & line, Limited Entry Sablefish longline, Limited Entry Fixed Gear DTL longline, Open Access Fixed Gear hook-&-line and the Directed P. halibut fishery using longline gear. In addition to year, the sector and gear should also be specified in both data.frames as a minimum. Other grouping variables can be included if desired.

Mortality rates are applied to the final mortality estimates. Mortality rates for hook & line injury assessments can be found in ./data/Constants.R.

NB: The final output (data.frame) will include a summary across all years for each stratum/grouping, in addition to a summary for each year; therefore nrow(output.df) > nrow(input.dfs).

IMPORTANT: This function assumes the project is stored in the following path: ~/observer/Input/PHLB/phlbbycatcher/. The mortality rates are stored in the data/Constants.R file which is called from this path. If different, change the path option of the function.

Note

For gear types that use viabilities (trawl, pot), use est_mort_viabs().

See also

Examples

if (FALSE) {
cshkl_injury <- est_mort_injury(cshkl_phlbiosp, csbhkl_totaldiscards,
  .mask = FALSE, .confs = NULL)
}