Estimate Unsampled Pacific halibut in the Catch Share Fishery

estimate_unsampld_cs is used to estimate the unsampled at-sea discards of Pacific halibut in the Catch Share fishery.

estimate_unsampld_cs(
  .l,
  .mask = TRUE,
  .rnd = FALSE,
  .decv = NULL,
  .quiet = TRUE,
  .nmfix = FALSE,
  .allstrata = FALSE,
  ...
)

Arguments

.l
a list, the first element (.l$meta) contains a data.frame called .l$metaconf_strata which simply lists the confidential strata (fishery, state, year, sector, gear, area, depth) and also a data.frame called .l$meta$strata which contains all the valid strata. The second element of the list should be a data.frame containing data from the Catch Share fishery to be used in estimating unsampled P. halibut bycatch.

.mask
Do you wish to mask the confidential strata? Default = TRUE.

.rnd
Only applies if .mask == TRUE. A logical indicating if the values should be rounded using round_df. Default = FALSE. See details.

.decv
Only applies if .rnd == TRUE. An integer indicating the number of decimal places to round, passed to base::round via round_df.

.quiet
Do you wish to suppress messages, mainly from dplyr? Default = TRUE.

.nmfix
Do you want to fix the names (i.e., namefix()) on output? Default = FALSE.

.allstrata
Do you want all strata generated? This is *not* the typical case and therefore, default is FALSE. The default will limit the output data set to only those strata found in .l$meta$strata (the most common case). However, in the case of monthly estimates for CS bottom trawl, RMONTH is not in .l$meta$strata, yet we need the data stratified at that level, thus set .allstrata = TRUE.

...
Names of columns (unquoted) in .l[[2]] (see above) by which the data will be grouped and summarized.

Value

Return a data.frame with an estimate of unsampled P. halibut bycatch and, optionally, confidential strata masked by an asterik.

Details

This function estimates the unsampled at-sea discards of P. halibut in the Catch Share fishery. Generally speaking it uses the sampled P. halibut:targeted landings ratio to estimate the weight of the unsampled P. halibut for each unsampled category (failed data, unsampled haul, unsampled mixed discards, and unsampled IFQ flatfish, IFQ mixed species) within a strata.

Examples

if (FALSE) {
# For example, to get an estimate of unsampled P. halibut by fishery, state, year, sector, gear, area, and depth (e.g., as one would for Catch Share bottom trawl).
est_unsampled_phlb_cs_btwl_masked <- estimate_unsampld_cs(.l = cs_btwl, fishery, state, RYEAR, sector, gear, area, depth, .mask = TRUE, .quietly = TRUE)

# Catch Share Monthly Estimates
est_monthly_cs_btwl_masked <- estimate_unsampld_cs(.l = cs_btwl, fishery, state, RYEAR, sector, gear, RMONTH, .mask = TRUE, .quietly = TRUE)
}