Create the P. halibut GEMM Product for Catch Shares Sectors

gemm_cs creates the P. halibut GEMM (Groundfish Expanded Multi-Year Mortality) product for Catch Shares sectors.

gemm_cs(
  sctr,
  gr,
  type = "nongroundfish",
  grouping = "Other nongroundfish",
  species = "Pacific Halibut",
  tot_landings_mt = 0
)

Arguments

sctr
A sector name that is found in the sector.args object stored in /Constants.R.

gr
A gear name that is found in the gear.args object stored in /Constants.R.

type
A grouping factor corresponding to the TYPE field in e.g., OBOrig_Proc. Should not need to change from the default "nongroundfish"

grouping
A grouping factor corresponding to the GROUPING field in e.g., OBOrig_Proc. Should not need to change from the default "Other nongroundfish"

species
A common species epithtet. Should not need to change from the default "Pacific Halibut"

tot_landings_mt
A default value (0) for total landings.

Value

Returns a data.frame of the formatted P. halibut GEMM product with correct GEMM sectors and data fields.

Details

The default values for type, grouping, species arguments are the appropriate case for what is needed in GEMM. This function **requires** a coverage object in /path_tables/coverage/ with an appropriate name. This file is built when the Catch Shares data is processed. This function **requires** a discard mortality object in path_data/<data.year>/calcs/ with an appropriate name. This file is built when the Catch Shares data is processed.

See also

Examples

if (FALSE) {
      gemm_btwl <- gemm_cs(sctr = "Catch Shares", gr = "Bottom Trawl")
      gemm_hkl <- gemm_cs(sctr = "Catch Shares", gr = "Hook & line")
      gemm_pot <- gemm_cs(sctr = "Catch Shares", gr = "Pot")
      gemm_hake <- gemm_cs(sctr = "Catch Shares", gr = "Midwater Hake")
      gemm_rock <- gemm_cs(sctr = "Catch Shares", gr = "Midwater Rock")
}