Assign Fishery Sectors to U.S. West Coast Fisheries Data

This low level helper function assigns a fishery sector to U.S. west coast fishery data, depending on the type of data supplied. See details

get.sector(X, sector, data.type)

Arguments

X
WCGOP pre-processed data set, typically OBOrig_Pre or FTOrig_Pre

sector
a character vector of length == 1 indicating a valid fishery sector name to be processed. See details for valid fishery sector names.

data.type
specifies the data source of X. See details for valid inputs.

Value

Returns the original data set with a new grouping variable, sector, which specifies the fishery sector.

Details

Fishery sectors differ from the initial FISHERY that is assigned upon data input. Multiple FISHERY codes could have the same sector assignment, e.g., sector = Nearshore will join all FISHERY = 'CA Nearshore','OR Blue-Black Rockfish Nearshore', or 'OR Blue-Black Rockfish' under the single Nearshore group.

Fishery sectors can be assigned to both the FOS Observer data (data.type = 'OB') or Pacific States Marine Fishery Commission fish ticket data (data.type = 'FT').

Valid fishery sectors include:

Catch Shares

Catch Shares EM

LE CA Halibut

Shoreside Hake

Midwater Hake

Midwater Rockfish

Nearshore

Limited Entry Trawl

Limited Entry Sablefish

LE Fixed Gear DTL

OA Fixed Gear

OA CA Halibut

Pink Shrimp

Directed P Halibut

Sea Cucumber

Ridgeback Prawn

See also

Examples

if (FALSE) {
le_sablefish_sector <- get.sector(X = OBOrig_Pre, sector = "Limited Entry Sablefish", data.type = "OB")
}