Create a Unique Haul Duration Object

haul_dur is used to create a unique haul duration object.

haul_dur(.df, ...)

Arguments

.df
a data.frame, typically from the NWFSC Observer Program. At minimum must contain the following data fields: DATATYPE, HAUL_ID, TRIP_ID, DRVID, HAUL_DURATION.

...
Names of columns (unquoted) in .df (see above) by which the data will be grouped and summarized.

Value

Return a data.frame with unique vessel-trip-haul-datatype haul durations.

Details

haul_dur creates a data.frame of all the unique haul durations for use in summarizing gear deployment times. In the context of phlbbycatcher, most frequently used for the IFQ tow time figure. Haul durations are given for each vessel-trip-haul-datatype and ... combination. Datatypes are A = Analysis Data, U = Unsampled Hauls, P = Partially Sampled Hauls.

NB: This function drops any cases where is.na(HAUL_DURATION) == TRUE, which happens when DATATYPE = "Trip Without Catch".

Examples

if (FALSE) {
haul_durations <- haul_dur(.df = df_with_haul_durations, fishery, state, RYEAR, sector, gear, area, depth)
}