In-Season A-SHOP P. halibut Mortality

insea_ashop_mort takes the in-season A-SHOP data and calculates the in-season mortality estimate and projects the mortality to the end of the year.

insea_ashop_mort(.asinsea, .ashist = ASOrig_Proc, .finalmonth, .finalday)

Arguments

.asinsea
A data.frame containing the A_SHOP in-season (typically 1 Jan- 15 Sept) PHLB bio-specimen data.

.ashist
A data.frame containing the historical A-SHOP catch data, including P. halibut bycatch.

.finalmonth
A numeric value (0-12) representing the final month to be *included* in each partial year's data. Defaults to 12.

.finalday
A numeric value (1-31) representing the final day of the month to be *included* in each partial year's data. Defaults to 31.

Value

Returns a data.frame of A-SHOP in-season mortality estimates.

See also

Examples

if (FALSE) {

}