In-Season Catch Shares P. halibut Mortality

insea_cs_mort takes the in-season Catch Shares viability data and calculates the in-season mortality estimate and projects the mortality to the end of the year.

insea_cs_mort(
  .viabsinsea,
  .obcatchinsea,
  .avepercent,
  .vas,
  .finalmonth,
  .finalday
)

Arguments

.viabsinsea
A data.frame containing the in-season (typically 1 Jan- 15 Sept) PHLB bio-specimen data, processed by insea_viabs_prep.

.obcatchinsea
A data.frame containing the in-season (typically 1 Jan- 15 Sept) observer catch data. #' @param .avepercent A data.frame containing historical percentage of catch in from observer data, obtained from insea_perc_cs_catch(). See details.

.vas
The in-season Vesssel Account System (lbs.) P. halibut bycatch north of 40 deg. 10 min. N. lat. (numeric, length == 1)

.finalmonth
A numeric value (0-12) representing the final month to be *included* in each partial year's data. Defaults to 12.

.finalday
A numeric value (1-31) representing the final day of the month to be *included* in each partial year's data. Defaults to 31.

Value

Returns a data.frame of Catch Shares in-season mortality estimates.

Details

insea_cs_mort Takes the in-season P. halibut bio-specimens that have been processed by insea_viabs_prep and the in-season observer catch data and uses these to estimate in-season mortality for the entire year. The unsampled amount is estimated from the Vessel Account System [VAS](https://www.webapps.nwfsc.noaa.gov/apex/ifq/f?p=155:1::::::) by subtracting the in-season sampled amount from the amount reported in the VAS - the difference being the extrapolated amount from the unsampled hauls (calculations done by VAS). The unsampled weight is distributed among gear types by multiplying the unsampled total (from VAS subtraction) by the proportion of unsampled categories (IFQFF, IFQM, NIFQ, UNST) among gear types in the in-season data. An important assumption is made: we assume that none of the unsampled weight is attributable to midwater trawl gear, because we almost never have unsampled categories in midwater trawl fisheries.

The unsampled weight is then distributed among viabilities, based on the proportion of viabilities found in the sampled P. halibut bio-specimens. We then multiply the sampled weight in each viability category by the viability-specific mortality rate. It is important to do this before adding the unsampled weight to the sampled weight because the unsampled weight from the VAS is already viability-adjusted (i.e, has been multiplied by the gear-viability appropriate mortality rate). Once the sampled weight is adjusted for viability by applying mortality rates, it is then added to the unsampled weight estimated from the VAS.

Now we have viability-adjusted P. halibut mortality, but only for the time-period included in the in-season data (typically 1st Jan - 15th Sept). We now need to project out to the remainder of the year (e.g, 16th Sept - 31st Dec). This is done by using the historical average proportion of mortality estimated for the period remaining in the year. In short, using previous years data, we estimate the average mortality for the period included in the in-season data (e.g., 1 Jan - 15 Sept) across years and the average mortality for the whole year (1 Jan - 31 Dec). The proportion (really, 1 - period average/annual average) is used to prorate the in-season mortality to the yet-to-be-fished time period (e.g., 16 Sept - 31 Dec). These calculations are produced by insea_perc_cs_catch().

See also

Examples

if (FALSE) {

}