In-Season Catch Share Percent P. halibut Catch

insea_perc_cs_catch calculates the average percentage of the P. halibut catch in the Catch Shares fleet that has occurred up to a given date in the year.

insea_perc_cs_catch(.finalmonth, .finalday, .inseayr, pth = NULL, ...)

Arguments

.finalmonth
A numeric value (0-12) representing the final month to be *included* in each partial year's data. Defaults to 12.

.finalday
A numeric value (1-31) representing the final day of the month to be *included* in each partial year's data. Defaults to 31.

.inseayr
A numeric year value (YYYY) representing the in-season year. This is used to *exclude* any erroneous in-season data that might pollute the historical data.

pth
A path to the historical P. halibut bio-specimen data (e.g., for 2020 in-season path should point to 2002-2019 data).

...
Arguments to ... will be used to dplyr::group_by() the catch summaries and should include a year grouping variable, at minimum.

Value

Returns a data.frame of historical percentages of Catch Shares bycatch up to a specified in-season date.

Details

insea_perc_cs_catch calculates the Catch Shares catch for the in-season period and then calculates the Catch Shares for the entire year. These two values are used to calculate the average percent of catch up to the final (latest) date of the in-season period (i.e., .finalmonth, .finalday). The average percentage is used to estimate the P. halibut bycatch mortality for the remainder of the year (i.e., the period *not* covered by the in-season data).

*IMPORTANT:* insea_perc_cs_catch assumes the presence of three objects in the global environment cs_btwl, cs_pot, cs_hkl. These objects are the output from stratify_4phlb() and contain the historical observer catch data for all years by gear type. See \scripts\Inseason_Calculations.Rmd to understand how these objects are created. Note that insea_perc_cs_catch uses the output from stratify_4phlb() to prepare the viabilities using prep_viabs().

See also

Examples