Estimate Legal Proportion of P. halibut Bycatch for CS

legal_prop_cs estimates the legal proportion of bycatch from P. halibut biological specimens (a.k.a., PHLB biospecimens; NWFSC FOS observer data) for use in mortality estimation and length frequency reporting. **NOTE:** This function only works for CS bottom trawl, pot, and midwater trawl. For CS hook and line, use phlbbycatcher::legal_prop_cshkl().

legal_prop_cs(
  sctr,
  ...,
  .mask = FALSE,
  .confs = NULL,
  nmfix = TRUE,
  path = path_data
)

Arguments

sctr
The sector to summarize legal proportions for. Determines loading of file.

...
The data fields to stratify the output.

.mask
Do you want to mask confidential data? Default is *FALSE*. If set to true, must supply a data.frame to the .confs argument.

.confs
(optional) A data.frame containing the unique strata/groups that whose accompanying data will be masked. Must be supplied if .mask = TRUE, otherwise, leave as NULL (default).

nmfix
Do you want to fix the names (i.e., namefix()) on output? Default = FALSE.

path
A path to the PHLB bio-specimens and the IPHC length-weight conversion table.

Value

Returns a data.frame that contains the legal proportion of discards for CS vessels using bottom or midwater trawl or pot.

Details

legal_prop_cs loads the PHLB biospecimen data, the total discard estimates from phlbbycather::estimate_unsampld_cs(), the viability and mortality estimates from phlbbycatcher::est_mort_viabs() and the International Pacific Halibut Commission (IPHC) length-weight conversion table; filters out EFP data, any <blank> viabilities, any viabilities where length is NA, and any BIO_SPECIMEN_METHOD not equal to 7, 9, or 12 (randomly sampled).

sctr options * cs_btwl = Catch Shares Bottom Trawl (non-EM) * cs_pot = Catch Shares Pot (non-EM) * cs_hake = Catch Shares Hake (includes both Shoreside and Midwater Hake) (non-EM) * cs_rock = Catch Shares Midwater Rockfish (non-EM)

See also

Examples

if (FALSE) {
csbtwl_legal <- legal_prop_cs(sctr = "cs_btwl", RYEAR, sector, gear, area, depth, path = path_data, .mask = TRUE, .confs = cs_btwl[[1]]$conf_strata)
}