Estimate Legal Proportion of P. halibut Bycatch for CS HooknLine Only

legal_prop_cshkl estimates the legal proportion of bycatch mortality from P. halibut biological specimens (a.k.a., PHLB biospecimens; NWFSC FOS observer data) for use in mortality estimation and length frequency reporting. **NOTE:** This function only works for CS hook and line. For other CS gear types, use phlbbycatcher::legal_prop_cs().

legal_prop_cshkl(
  sctr = "cs_hkl",
  ...,
  .mask = FALSE,
  .confs = cs_hkl[[1]]$conf_strata,
  nmfix = FALSE,
  path = path_data
)

Arguments

sctr
The sector to summarize legal proportions for, must be cs_hkl. Determines loading of appropriate files.

...
The data fields to stratify the output.

.mask
Do you want to mask confidential data? Default is *FALSE*. If set to true, must supply a data.frame to the .confs argument.

.confs
(optional) A data.frame containing the unique strata/groups that whose accompanying data will be masked. Must be supplied if .mask = TRUE, otherwise, leave as NULL (default).

nmfix
Do you want to fix the names (i.e., namefix()) on output? Default = FALSE.

path
A path to the PHLB bio-specimens and the IPHC length-weight conversion table. Default is path_data, which is defined in Setup.Rmd.

Value

Returns a data.frame that contains the legal proportion of discards for CS vessels using hook & line gear only.

Details

legal_prop_cshkl loads the PHLB biospecimen data, the total discard estimates from phlbbycather::estimate_unsampld_cs(), the injury and mortality estimates from phlbbycatcher::est_mort_viabs() and phlbbycatcher""est_mort_viabs_cs_hkl1119() and the International Pacific Halibut Commission (IPHC) length-weight conversion table; filters out EFP data, any <blank> viabilities, any viabilities where length is NA, and any BIO_SPECIMEN_METHOD not equal to 7, 9, or 12 (randomly sampled).

#' For the period 2011-2017, estimates are made using a constant discard mortality rate (DMR) of 0.16. For the period 2018-2020, estimates are made using both a constant DMR (0.16) and the injury method, similar to the viability method used in bottom trawl. Injuries are categorized as minor, moderate, serious, or dead. Mortality rates associated with each injury category are defined in /scripts/Constants.R and based on [Trumble et al. (1998)](https://doi.org/10.1577/1548-8675(2000)020<0931:EODMRF>2.0.CO;2)

sctr options * cs_hkl = Catch Shares Hook & Line (non-EM) - currently this is the only option available for this function.

See also

Examples