Match Trip Data to P. Halibut Biosamples

match_trip_phlb matches the trip data from observer data (e.g., OBOrig_...) with the P. halibut biosample data from the data base.

match_trip_phlb(.df, as = FALSE, .v = FALSE, ...)

Arguments

.df
A data.frame of NWFSC observer data.

as
A logical indicating if these are A-SHOP (TRUE) or WCGOP (FALSE) length frequencies. Only works when .v = FALSE, otherwise has no effect.

.v
A logical indicating if these are to be for viablities (TRUE) or length frequency summaries (default = FALSE).

...
Any other fields (besides sector, gear) to be added to the biosamples data output.

Value

Returns a data.frame of P. halibut biosamples with trips matched and new fields (at minimum sector, gear) added from OBOrig_...

Details

See also

Examples

if (FALSE) {
cs_btwl_lfs <- match_trip_phlb( .df = cs_btwl$out, .v = FALSE, as = FALSE, fishery, state, RYEAR, sector, gear, area, depth)
}