Prepare Viabilities for Use in P. halibut Mortality Estimation

prep_viabs prepares the viability data from P. halibut biological specimens (a.k.a., PHLB biospecimens; NWFSC FOS observer data) for use in mortality estimation and length frequency reporting.

prep_viabs(
  .df,
  .strata = NULL,
  .mask = FALSE,
  .confs = NULL,
  nmfix = FALSE,
  path = "~/observer/Input/PHLB/phlbbycatcher/data",
  ...
)

Arguments

.df
A data.frame containing haul-level information to be merged with the PHLB biospecimen data. At minimum, .df must contain: TRIP_ID, HAUL_ID, HAUL_NUM, AVG_LAT, AVG_DEPTH. Strongly suggested to also contain: sector, gear, RYEAR, which will be supplied via ... .

.strata
A data.frame containing the strata contained in the .df. This is a required filed and can be retrieved from the meta object created by phlbbycather::stratify_4phlb(). Used to limit the output to the strata contained in .df.

.mask
Do you want to mask confidential data? Default is *FALSE*. If set to true, must supply a data.frame to the .confs argument.

.confs
(optional) A data.frame containing the unique strata/groups that whose accompanying data will be masked. Must be supplied if .mask = TRUE, otherwise, leave as NULL (default).

nmfix
Do you want to fix the names (i.e., namefix()) on output? Default = FALSE.

path
A path to the PHLB bio-specimens and the IPHC length-weight conversion table.

...
Names of columns (unquoted) in .df by which the data will be grouped and summarized. Strongly suggest a minimum of : sector, gear, RYEAR. Could also include any other grouping variables found in the data, e.g., area, depth.

Value

Returns a data.frame that contains the grouping/stratification variables (from ...), and counts and weights of P. halibut in each viability category (as columns).

Details

prep_viabs loads the PHLB biospecimen data and the International Pacific Halibut Commission (IPHC) length-weight conversion table; filters out EFP data, any <blank> viabilities, any viabilities where length is NA, and any BIO_SPECIMEN_METHOD not equal to 7 or 9 (randomly sampled). The function matches the biospecimens/viability data with trip information from .df and ..., thus dropping any viabilities that do not match the trip information. Finally, prep_viabs uses the IPHC length-weight conversion table to convert the lengths (in PHLB biospecimens) to weights for each viability category.

prep_viabs should only be necessary for those sector-gear combinations that utilize viabilities in the estimation of P. halibut mortality. Those sector-gear combinations include: IFQ bottom trawl, IFQ pot, Limited Entry Sablefish longline, Limited Entry Fixed Gear DTL longline, Open Access Fixed Gear hook-&-line and the Directed P. halibut fishery using longline gear. Thus specifying the sector and gear in the ... argument is absolutely necessary.

Note that viability categories differ depending on gear type. For bottom trawl and pot gears, viability categories are excellent, poor, and dead. For hook-&-line gears, injury categories are minor, moderate, serious, and dead. See the NWFSC Observer Manual for definitions and determinations of P. halibut viability/injury. prep_viabs treats the different viabilities appropriately depending on gear type.

See also

Examples

if (FALSE) {
csbtwl_viabs <- prep_viabs(
  csbtwl, state, fishery, RYEAR, sector,
  gear, area, depth
)
}