Stratify and Subset Data for Pacific halibut Bycatch Estimates

A wrapper function to subset and stratify the NWFSC West Coast Groundfish Observer Program data in preparation for estimating Pacific halibut bycatch in the Catch Share and non-Catch Share fisheries. For data from At-sea Pacific hake fisheries, see the estimate_ashop_phlb function.

stratify_4phlb(
  df,
  sct = NULL,
  spc = SPC,
  gr = NULL,
  a_lat = NULL,
  a_names = NULL,
  d_dpth = NULL,
  d_names = NULL,
  state = FALSE,
  notes = NULL
)

Arguments

df
WCGOP pre-processed data set, typically OBOrig_Pre (but note that OBOrig_Proc will likely work too). If OBOrig_Pre is used, you *must* have the SPC file loaded in the global environment. See details.

sct
a character string indicating the specific fishery sector which the data will be subset. See subset_sector for valid options.

spc
this is the current FOS SPC file which adds important fields to the OBOrig_Pre object.

gr
a character string indicating the specific gear type which the data will be subset. See subset_gear for valid options.

a_lat
a numeric vector of latitudes, in decimal format.

a_names
a character vector of labels (minimum 2) to be associated with each latitudinal break. The length of a_names must be equal to length(a_lat) + 1.

d_dpth
a numeric vector of depths.

d_names
a character vector of labels (minimum 2) to be associated with each depth bin. The length of d_names must be equal to length(d_dpth) + 1.

state
a logical indicating if the data should be stratified by state. Default (FALSE) produces a state column with WA,OR,CA for each row in the data.

notes
Any notes associated with the run. Will be appended to the meta-data. (character vector, default = NULL)

...
other arguments. See base::cut for finer control of area stratification.

Value

Returns a list comprised of:

out: the subsetted data (data.frame) with state, area, and depth strata applied.

meta: a list of meta-data, as produced either by this function or the add_meta function.

Details

WCGOP data set (typically , OBOrig_Pre). If OBOrig_Pre is passed then OB.processing is run on the object before being acted upon. OBOrig_Pre requires that the most recent SPC file is loaded in the global environment. If using OBOrig_Proc, the data must have the following data fields:

sector

gear - Note: unlike all other headers, this header is case sensitive.

AVG_LAT

AVG_DEPTH

R_STATE

See also

Examples

if (FALSE) {
cs_btwl <- stratify_4phlb(df = OBOrig_Proc, sct = "Catch Shares", gr = "Bottom Trawl", a_lat = c(40.1667, 46.88825), a_names = c(s4010, ptch4010, nptch), d_dpth = c(60), d_names = c("0-60", ">60"), state = FALSE, notes = NULL)

distinct(cs_btwl$out, fishery, state, sector, gear, area, depth)
}