Catch Share Coverage Tables

summrize_cs_coverage creates the Catch Shares coverage tables.

summarize_cs_coverage(
  .df,
  .mask = TRUE,
  .confs = NULL,
  .rnd = TRUE,
  .decv = 2,
  .quiet = TRUE,
  .nmfix = FALSE,
  ...
)

Arguments

.df
A data.frame, see details.

.mask
Do you wish to mask confidential data with an * ? Default = TRUE.

.confs
(optional) A data.frame containing the unique strata/groups that whose accompanying data will be masked. Must be supplied if .mask = TRUE, otherwise, leave as NULL (default).

.decv
Only applies if rnd == TRUE. An integer indicating the number of decimal places to round (default = 2), passed to base::round via round_df.

.quiet
Do you wish to suppress messages, mainly from dplyr? Default = TRUE.

.nmfix
Do you want to fix the names (i.e., namefix()) on output? Default = FALSE.

...
Names of columns (unquoted) in .df by which the data will be grouped, i.e., the strata.

rnd
A logical indicating if the values should be rounded using round_df. Default = FALSE. See details.

Value

Returns a non-confidential (default) Catch Share observer coverage table as a data.frame, grouped according to ....

Details

This function creates the Catch Share coverage tables at the grouping/strata specified in .... By default, the function automatically replaces confidential data with an asterisk, when the number of vessels in a given strata are less than 3. To prevent masking, set .mask = FALSE. Note that this function only returns observer data and does not contain data from fish tickets.

Examples

if (FALSE) {
cov_cs_btwl <- summarize_cs_coverage(
  cs_btwl, state, fishery, RYEAR,
  sector, gear, area, depth
)
}